pyproject.toml
[tool.poetry]
name = "fastapi-poetry-template"
version = "0.1.0"
description = ""
authors = ["Yuchan Han <serltretu24@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
pydantic-settings = "^2.6.1"
sentry-sdk = {extras = ["fastapi"], version = "^2.19.0"}
uvicorn = "^0.32.1"
fastapi-restful = "^0.6.0"
typing-inspect = "^0.9.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
